requirements.txt
django == 3.1.2
pillow
django-admin-interface
django-rest-framework